pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Dmitry_Perexozhev <perexozhev99@gmail.com>"]
readme = "README.md"
packages = [
    {include = "task_manager"}
]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
dj-database-url = "^2.2.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
uvicorn = "^0.31.1"
django-bootstrap5 = "^24.3"
django-filter = "^24.3"
rollbar = "^1.0.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
